refactor: update hsaf
